Move OGR processing provider specific methods to GdalUtils
