+++ python/plugins/processing/algs/gdal/ClipByMask.py
@@ -45,7 +45,6 @@ from processing.algs.gdal.GdalAlgorithm import GdalAlgorithm
 from processing.algs.gdal.GdalUtils import GdalUtils
 
 from processing.tools import dataobjects
-from processing.tools.vector import ogrConnectionString
 
 pluginPath = os.path.split(os.path.split(os.path.dirname(__file__))[0])[0]
 
@@ -109,7 +108,7 @@ class ClipByMask(GdalAlgorithm):
         mask = self.getParameterValue(self.MASK)
         context = dataobjects.createContext()
         maskLayer = QgsProcessingUtils.mapLayerFromString(self.getParameterValue(self.MASK), context)
-        ogrMask = ogrConnectionString(mask, context)[1:-1]
+        ogrMask = GdalUtils.ogrConnectionString(mask, context)[1:-1]
         noData = self.getParameterValue(self.NO_DATA)
         opts = self.getParameterValue(self.OPTIONS)
