+++ python/plugins/processing/algs/gdal/ogr2ogr.py
@@ -37,7 +37,6 @@ from processing.algs.gdal.GdalAlgorithm import GdalAlgorithm
 from processing.algs.gdal.GdalUtils import GdalUtils
 
 from processing.tools.system import isWindows
-from processing.tools.vector import ogrConnectionString, ogrLayerName
 
 
 FORMATS = [
@@ -124,7 +123,7 @@ class Ogr2Ogr(GdalAlgorithm):
 
     def getConsoleCommands(self, parameters, context, feedback):
         inLayer = self.getParameterValue(self.INPUT_LAYER)
-        ogrLayer = ogrConnectionString(inLayer, context)[1:-1]
+        ogrLayer = GdalUtils.ogrConnectionString(inLayer, context)[1:-1]
 
         output = self.getOutputFromName(self.OUTPUT_LAYER)
         outFile = output.value
@@ -136,7 +135,7 @@ class Ogr2Ogr(GdalAlgorithm):
             outFile += ext
             output.value = outFile
 
-        output = ogrConnectionString(outFile, context)
+        output = GdalUtils.ogrConnectionString(outFile, context)
         options = str(self.getParameterValue(self.OPTIONS))
 
         if outFormat == 'SQLite' and os.path.isfile(output):
@@ -151,7 +150,7 @@ class Ogr2Ogr(GdalAlgorithm):
 
         arguments.append(output)
         arguments.append(ogrLayer)
-        arguments.append(ogrLayerName(inLayer))
+        arguments.append(GdalUtils.ogrLayerName(inLayer))
 
         commands = []
         if isWindows():
