+++ python/plugins/processing/algs/gdal/ogr2ogrclip.py
@@ -35,7 +35,6 @@ from processing.algs.gdal.GdalUtils import GdalUtils
 
 from processing.tools import dataobjects
 from processing.tools.system import isWindows
-from processing.tools.vector import ogrConnectionString, ogrLayerName
 
 
 class Ogr2OgrClip(GdalAlgorithm):
@@ -69,28 +68,28 @@ class Ogr2OgrClip(GdalAlgorithm):
 
     def getConsoleCommands(self, parameters, context, feedback):
         inLayer = self.getParameterValue(self.INPUT_LAYER)
-        ogrLayer = ogrConnectionString(inLayer, context)[1:-1]
+        ogrLayer = GdalUtils.ogrConnectionString(inLayer, context)[1:-1]
         clipLayer = self.getParameterValue(self.CLIP_LAYER)
-        ogrClipLayer = ogrConnectionString(clipLayer, context)[1:-1]
+        ogrClipLayer = GdalUtils.ogrConnectionString(clipLayer, context)[1:-1]
 
         output = self.getOutputFromName(self.OUTPUT_LAYER)
         outFile = output.value
 
-        output = ogrConnectionString(outFile, context)
+        output = GdalUtils.ogrConnectionString(outFile, context)
         options = str(self.getParameterValue(self.OPTIONS))
 
         arguments = []
         arguments.append('-clipsrc')
         arguments.append(ogrClipLayer)
         arguments.append("-clipsrclayer")
-        arguments.append(ogrLayerName(clipLayer))
+        arguments.append(GdalUtils.ogrLayerName(clipLayer))
 
         if options is not None and len(options.strip()) > 0:
             arguments.append(options)
 
         arguments.append(output)
         arguments.append(ogrLayer)
-        arguments.append(ogrLayerName(inLayer))
+        arguments.append(GdalUtils.ogrLayerName(inLayer))
 
         commands = []
         if isWindows():
