+++ python/plugins/processing/algs/gdal/ogr2ogrclipextent.py
@@ -35,7 +35,6 @@ from processing.algs.gdal.GdalAlgorithm import GdalAlgorithm
 from processing.algs.gdal.GdalUtils import GdalUtils
 
 from processing.tools.system import isWindows
-from processing.tools.vector import ogrConnectionString, ogrLayerName
 
 
 class Ogr2OgrClipExtent(GdalAlgorithm):
@@ -69,7 +68,7 @@ class Ogr2OgrClipExtent(GdalAlgorithm):
 
     def getConsoleCommands(self, parameters, context, feedback):
         inLayer = self.getParameterValue(self.INPUT_LAYER)
-        ogrLayer = ogrConnectionString(inLayer, context)[1:-1]
+        ogrLayer = GdalUtils.ogrConnectionString(inLayer, context)[1:-1]
         clipExtent = self.getParameterValue(self.CLIP_EXTENT)
         if not clipExtent:
             clipExtent = QgsProcessingUtils.combineLayerExtents([inLayer])
@@ -77,7 +76,7 @@ class Ogr2OgrClipExtent(GdalAlgorithm):
         output = self.getOutputFromName(self.OUTPUT_LAYER)
         outFile = output.value
 
-        output = ogrConnectionString(outFile, context)
+        output = GdalUtils.ogrConnectionString(outFile, context)
         options = str(self.getParameterValue(self.OPTIONS))
 
         arguments = []
@@ -94,7 +93,7 @@ class Ogr2OgrClipExtent(GdalAlgorithm):
 
         arguments.append(output)
         arguments.append(ogrLayer)
-        arguments.append(ogrLayerName(inLayer))
+        arguments.append(GdalUtils.ogrLayerName(inLayer))
 
         commands = []
         if isWindows():
