+++ python/plugins/processing/algs/gdal/ogr2ogrdissolve.py
@@ -36,7 +36,6 @@ from processing.algs.gdal.GdalUtils import GdalUtils
 
 from processing.tools import dataobjects
 from processing.tools.system import isWindows
-from processing.tools.vector import ogrConnectionString, ogrLayerName
 
 
 class Ogr2OgrDissolve(GdalAlgorithm):
@@ -103,13 +102,13 @@ class Ogr2OgrDissolve(GdalAlgorithm):
         statsatt = self.getParameterValue(self.STATSATT)
         options = self.getParameterValue(self.OPTIONS)
 
-        ogrLayer = ogrConnectionString(inLayer, context)[1:-1]
-        layername = ogrLayerName(inLayer)
+        ogrLayer = GdalUtils.ogrConnectionString(inLayer, context)[1:-1]
+        layername = GdalUtils.ogrLayerName(inLayer)
 
         output = self.getOutputFromName(self.OUTPUT_LAYER)
         outFile = output.value
 
-        output = ogrConnectionString(outFile, context)
+        output = GdalUtils.ogrConnectionString(outFile, context)
 
         arguments = []
         arguments.append(output)
