+++ python/plugins/processing/algs/gdal/ogr2ogrtabletopostgislist.py
@@ -39,7 +39,6 @@ from processing.algs.gdal.GdalUtils import GdalUtils
 
 from processing.tools.postgis import uri_from_name, GeoDB
 from processing.tools.system import isWindows
-from processing.tools.vector import ogrConnectionString, ogrLayerName
 
 
 class Ogr2OgrTableToPostGisList(GdalAlgorithm):
@@ -141,7 +140,7 @@ class Ogr2OgrTableToPostGisList(GdalAlgorithm):
             uri = GeoDB(uri=uri).uri
 
         inLayer = self.getParameterValue(self.INPUT_LAYER)
-        ogrLayer = ogrConnectionString(inLayer, context)[1:-1]
+        ogrLayer = GdalUtils.ogrConnectionString(inLayer, context)[1:-1]
         shapeEncoding = self.getParameterValue(self.SHAPE_ENCODING)
         schema = str(self.getParameterValue(self.SCHEMA))
         table = str(self.getParameterValue(self.TABLE))
@@ -176,7 +175,7 @@ class Ogr2OgrTableToPostGisList(GdalAlgorithm):
         arguments.append('"')
         arguments.append(ogrLayer)
         arguments.append('-nlt NONE')
-        arguments.append(ogrLayerName(inLayer))
+        arguments.append(GdalUtils.ogrLayerName(inLayer))
         if launder:
             arguments.append(launderstring)
         if append:
@@ -190,7 +189,7 @@ class Ogr2OgrTableToPostGisList(GdalAlgorithm):
         elif primary_key is not None:
             arguments.append("-lco FID=" + primary_key)
         if len(table) == 0:
-            table = ogrLayerName(inLayer).lower()
+            table = GdalUtils.ogrLayerName(inLayer).lower()
         if schema:
             table = '{}.{}'.format(schema, table)
         arguments.append('-nln')
