+++ python/plugins/processing/tests/GdalAlgorithmsTest.py
@@ -27,14 +27,19 @@ __revision__ = ':%H$'
 
 import AlgorithmsTestBase
 from processing.algs.gdal.ogr2ogrtopostgis import Ogr2OgrToPostGis
+from processing.algs.gdal.GdalUtils import GdalUtils
 from qgis.core import QgsProcessingContext
 import nose2
+import os
 import shutil
+import tempfile
 
 from qgis.testing import (
     start_app,
     unittest
 )
+
+testDataPath = os.path.join(os.path.dirname(__file__), 'testdata')
 
 
 class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
@@ -53,6 +58,45 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
 
     def test_definition_file(self):
         return 'gdal_algorithm_tests.yaml'
+
+    def testOgrLayerNameExtraction(self):
+        outdir = tempfile.mkdtemp()
+        self.cleanup_paths.append(outdir)
+
+        def _copyFile(dst):
+            shutil.copyfile(os.path.join(testDataPath, 'custom', 'grass7', 'weighted.csv'), dst)
+
+        # OGR provider - single layer
+        _copyFile(os.path.join(outdir, 'a.csv'))
+        name = GdalUtils.ogrLayerName(outdir)
+        self.assertEqual(name, 'a')
+
+        # OGR provider - multiple layers
+        _copyFile(os.path.join(outdir, 'b.csv'))
+        name1 = GdalUtils.ogrLayerName(outdir + '|layerid=0')
+        name2 = GdalUtils.ogrLayerName(outdir + '|layerid=1')
+        self.assertEqual(sorted([name1, name2]), ['a', 'b'])
+
+        name = GdalUtils.ogrLayerName(outdir + '|layerid=2')
+        self.assertIsNone(name)
+
+        # OGR provider - layername takes precedence
+        name = GdalUtils.ogrLayerName(outdir + '|layername=f')
+        self.assertEqual(name, 'f')
+
+        name = GdalUtils.ogrLayerName(outdir + '|layerid=0|layername=f')
+        self.assertEqual(name, 'f')
+
+        name = GdalUtils.ogrLayerName(outdir + '|layername=f|layerid=0')
+        self.assertEqual(name, 'f')
+
+        # SQLiite provider
+        name = GdalUtils.ogrLayerName('dbname=\'/tmp/x.sqlite\' table="t" (geometry) sql=')
+        self.assertEqual(name, 't')
+
+        # PostgreSQL provider
+        name = GdalUtils.ogrLayerName('port=5493 sslmode=disable key=\'edge_id\' srid=0 type=LineString table="city_data"."edge" (geom) sql=')
+        self.assertEqual(name, 'city_data.edge')
 
 
 class TestGdalOgr2OgrToPostgis(unittest.TestCase):
